Backend/requirements.txt
google-generativeai
yfinance
pandas
requests
Flask
Flask-Cors
gunicorn
nsepy